Python/LinkedIn/requirements.txt
streamlit
feedparser
pyyaml
openai
python-dotenv